Move conftest.py from repository root to tests directory (#17718)
