--- conftest.py
@@ -1,520 +1,0 @@
-import json
-import os
-import posixpath
-import re
-import shutil
-import subprocess
-import sys
-import threading
-import time
-from collections import defaultdict
-from dataclasses import dataclass
-from pathlib import Path
-
-import pytest
-
-from mlflow.environment_variables import _MLFLOW_TESTING, MLFLOW_TRACKING_URI
-from mlflow.utils.os import is_windows
-from mlflow.version import IS_TRACING_SDK_ONLY, VERSION
-
-
-def pytest_addoption(parser):
-    parser.addoption(
-        "--requires-ssh",
-        action="store_true",
-        dest="requires_ssh",
-        default=False,
-        help="Run tests decorated with 'requires_ssh' annotation. "
-        "These tests require keys to be configured locally "
-        "for SSH authentication.",
-    )
-    parser.addoption(
-        "--ignore-flavors",
-        action="store_true",
-        dest="ignore_flavors",
-        default=False,
-        help="Ignore tests for model flavors.",
-    )
-    parser.addoption(
-        "--splits",
-        default=None,
-        type=int,
-        help="The number of groups to split tests into.",
-    )
-    parser.addoption(
-        "--group",
-        default=None,
-        type=int,
-        help="The group of tests to run.",
-    )
-    parser.addoption(
-        "--serve-wheel",
-        action="store_true",
-        default=os.getenv("CI", "false").lower() == "true",
-        help="Serve a wheel for the dev version of MLflow. True by default in CI, False otherwise.",
-    )
-
-
-def pytest_configure(config):
-    config.addinivalue_line("markers", "requires_ssh")
-    config.addinivalue_line("markers", "notrackingurimock")
-    config.addinivalue_line("markers", "allow_infer_pip_requirements_fallback")
-    config.addinivalue_line(
-        "markers", "do_not_disable_new_import_hook_firing_if_module_already_exists"
-    )
-    config.addinivalue_line("markers", "classification")
-    config.addinivalue_line("markers", "no_mock_requests_get")
-
-    labels = fetch_pr_labels() or []
-    if "fail-fast" in labels:
-        config.option.maxfail = 1
-
-
-@pytest.hookimpl(tryfirst=True)
-def pytest_cmdline_main(config: pytest.Config):
-    if not_exists := [p for p in config.getoption("ignore") or [] if not os.path.exists(p)]:
-        raise pytest.UsageError(f"The following paths are ignored but do not exist: {not_exists}")
-
-    group = config.getoption("group")
-    splits = config.getoption("splits")
-
-    if splits is None and group is None:
-        return None
-
-    if splits and group is None:
-        raise pytest.UsageError("`--group` is required")
-
-    if group and splits is None:
-        raise pytest.UsageError("`--splits` is required")
-
-    if splits < 0:
-        raise pytest.UsageError("`--splits` must be >= 1")
-
-    if group < 1 or group > splits:
-        raise pytest.UsageError("`--group` must be between 1 and {splits}")
-
-    return None
-
-
-@dataclass
-class TestResult:
-    path: Path
-    test_name: str
-    execution_time: float
-
-
-_test_results: list[TestResult] = []
-
-
-def pytest_sessionstart(session):
-    # Clear duration tracking state at the start of each session
-    _test_results.clear()
-
-    if IS_TRACING_SDK_ONLY:
-        return
-
-    import click
-
-    if uri := MLFLOW_TRACKING_URI.get():
-        click.echo(
-            click.style(
-                (
-                    f"Environment variable {MLFLOW_TRACKING_URI} is set to {uri!r}, "
-                    "which may interfere with tests."
-                ),
-                fg="red",
-            )
-        )
-
-
-def to_md_table(rows: list[list[str]]) -> str:
-    if not rows:
-        return ""
-    n = max(len(r) for r in rows)
-    rows = [r + [""] * (n - len(r)) for r in rows]
-
-    # Calculate column widths
-    widths = [max(len(row[i]) for row in rows) for i in range(n)]
-
-    def esc(s: str) -> str:
-        return s.replace("|", r"\|").replace("\n", "<br>")
-
-    # Format rows with proper padding
-    def format_row(row: list[str]) -> str:
-        cells = [esc(cell).ljust(width) for cell, width in zip(row, widths)]
-        return "| " + " | ".join(cells) + " |"
-
-    header = format_row(rows[0])
-    sep = "| " + " | ".join(["-" * w for w in widths]) + " |"
-    body = [format_row(row) for row in rows[1:]]
-
-    return "\n".join([header, sep, *body])
-
-
-def generate_duration_stats() -> str:
-    """Generate per-file duration statistics as markdown table."""
-    if not _test_results:
-        return ""
-
-    # Group results by file path
-    file_groups: defaultdict[Path, list[float]] = defaultdict(list)
-    for result in _test_results:
-        file_groups[result.path].append(result.execution_time)
-
-    rows = []
-    for path, test_times in file_groups.items():
-        rel_path = path.relative_to(Path.cwd()).as_posix()
-        total_dur = sum(test_times)
-        if total_dur < 1.0:
-            # Ignore files with total duration < 1s
-            continue
-        test_count = len(test_times)
-        min_test = min(test_times)
-        max_test = max(test_times)
-        avg_test = sum(test_times) / len(test_times)
-
-        rows.append((rel_path, total_dur, test_count, min_test, max_test, avg_test))
-
-    rows.sort(key=lambda r: r[1], reverse=True)
-
-    if not rows:
-        return ""
-
-    # Prepare data for markdown table (headers + data rows)
-    table_rows = [["Rank", "File", "Duration", "Tests", "Min", "Max", "Avg"]]
-    for idx, (path, dur, count, min_, max_, avg_) in enumerate(rows, 1):
-        table_rows.append(
-            [
-                str(idx),
-                f"`{path}`",
-                f"{dur:.2f}s",
-                str(count),
-                f"{min_:.3f}s",
-                f"{max_:.3f}s",
-                f"{avg_:.3f}s",
-            ]
-        )
-
-    return to_md_table(table_rows)
-
-
-@pytest.hookimpl(hookwrapper=True)
-def pytest_runtest_protocol(item: pytest.Item, nextitem: pytest.Item | None):
-    start = time.perf_counter()
-    yield  # This includes setup + call + teardown
-    duration = time.perf_counter() - start
-    _test_results.append(TestResult(path=item.path, test_name=item.name, execution_time=duration))
-
-
-def pytest_runtest_setup(item):
-    markers = [mark.name for mark in item.iter_markers()]
-    if "requires_ssh" in markers and not item.config.getoption("--requires-ssh"):
-        pytest.skip("use `--requires-ssh` to run this test")
-
-
-def fetch_pr_labels():
-    """
-    Returns the labels associated with the current pull request.
-    """
-    if "GITHUB_ACTIONS" not in os.environ:
-        return None
-
-    if os.environ.get("GITHUB_EVENT_NAME") != "pull_request":
-        return None
-
-    with open(os.environ["GITHUB_EVENT_PATH"]) as f:
-        pr_data = json.load(f)
-        return [label["name"] for label in pr_data["pull_request"]["labels"]]
-
-
-@pytest.hookimpl(hookwrapper=True)
-def pytest_report_teststatus(report, config):
-    outcome = yield
-    if report.when == "call":
-        try:
-            import psutil
-        except ImportError:
-            return
-
-        (*rest, result) = outcome.get_result()
-        mem = psutil.virtual_memory()
-        mem_used = mem.used / 1024**3
-        mem_total = mem.total / 1024**3
-
-        disk = psutil.disk_usage("/")
-        disk_used = disk.used / 1024**3
-        disk_total = disk.total / 1024**3
-        outcome.force_result(
-            (
-                *rest,
-                (
-                    f"{result} | "
-                    f"MEM {mem_used:.1f}/{mem_total:.1f} GB | "
-                    f"DISK {disk_used:.1f}/{disk_total:.1f} GB"
-                ),
-            )
-        )
-
-
-@pytest.hookimpl(hookwrapper=True)
-def pytest_ignore_collect(collection_path, config):
-    outcome = yield
-    if not outcome.get_result() and config.getoption("ignore_flavors"):
-        # If not ignored by the default hook and `--ignore-flavors` specified
-
-        # Ignored files and directories must be included in dev/run-python-flavor-tests.sh
-        model_flavors = [
-            # Tests of flavor modules.
-            "tests/ag2",
-            "tests/agno",
-            "tests/anthropic",
-            "tests/autogen",
-            "tests/azureml",
-            "tests/bedrock",
-            "tests/catboost",
-            "tests/crewai",
-            "tests/diviner",
-            "tests/dspy",
-            "tests/gemini",
-            "tests/groq",
-            "tests/h2o",
-            "tests/johnsnowlabs",
-            "tests/keras",
-            "tests/keras_core",
-            "tests/llama_index",
-            "tests/langchain",
-            "tests/langgraph",
-            "tests/lightgbm",
-            "tests/litellm",
-            "tests/mistral",
-            "tests/models",
-            "tests/onnx",
-            "tests/openai",
-            "tests/paddle",
-            "tests/pmdarima",
-            "tests/promptflow",
-            "tests/prophet",
-            "tests/pydantic_ai",
-            "tests/pyfunc",
-            "tests/pytorch",
-            "tests/strands",
-            "tests/haystack",
-            "tests/semantic_kernel",
-            "tests/sentence_transformers",
-            "tests/shap",
-            "tests/sklearn",
-            "tests/smolagents",
-            "tests/spacy",
-            "tests/spark",
-            "tests/statsmodels",
-            "tests/tensorflow",
-            "tests/transformers",
-            "tests/xgboost",
-            # Lazy loading test.
-            "tests/test_mlflow_lazily_imports_ml_packages.py",
-            # This test is included here because it imports many big libraries like tf, keras, etc.
-            "tests/tracking/fluent/test_fluent_autolog.py",
-            # Cross flavor autologging related tests.
-            "tests/autologging/test_autologging_safety_unit.py",
-            "tests/autologging/test_autologging_behaviors_unit.py",
-            "tests/autologging/test_autologging_behaviors_integration.py",
-            "tests/autologging/test_autologging_utils.py",
-            "tests/autologging/test_training_session.py",
-        ]
-
-        relpath = os.path.relpath(str(collection_path))
-        relpath = relpath.replace(os.sep, posixpath.sep)  # for Windows
-
-        if relpath in model_flavors:
-            outcome.force_result(True)
-
-
-@pytest.hookimpl(trylast=True)
-def pytest_collection_modifyitems(session, config, items):
-    # Executing `tests.server.test_prometheus_exporter` after `tests.server.test_handlers`
-    # results in an error because Flask >= 2.2.0 doesn't allow calling setup method such as
-    # `before_request` on the application after the first request. To avoid this issue,
-    # execute `tests.server.test_prometheus_exporter` first by reordering the test items.
-    items.sort(key=lambda item: item.module.__name__ != "tests.server.test_prometheus_exporter")
-
-    # Select the tests to run based on the group and splits
-    if (splits := config.getoption("--splits")) and (group := config.getoption("--group")):
-        items[:] = items[(group - 1) :: splits]
-
-
-@pytest.hookimpl(hookwrapper=True)
-def pytest_terminal_summary(terminalreporter, exitstatus, config):
-    yield
-
-    # Display per-file durations
-    if duration_stats := generate_duration_stats():
-        terminalreporter.write("\n")
-        header = "per-file durations (sorted)"
-        terminalreporter.write_sep("=", header)
-        terminalreporter.write(f"::group::{header}\n\n")
-        terminalreporter.write(duration_stats)
-        terminalreporter.write("\n\n::endgroup::\n")
-        terminalreporter.write("\n")
-
-    # If there are failed tests, display a command to run them
-    failed_test_reports = terminalreporter.stats.get("failed", [])
-    if failed_test_reports:
-        if len(failed_test_reports) <= 30:
-            ids = [repr(report.nodeid) for report in failed_test_reports]
-        else:
-            # Use dict.fromkeys to preserve the order
-            ids = list(dict.fromkeys(report.fspath for report in failed_test_reports))
-        terminalreporter.section("command to run failed tests")
-        terminalreporter.write(" ".join(["pytest"] + ids))
-        terminalreporter.write("\n" * 2)
-
-        if summary_path := os.environ.get("GITHUB_STEP_SUMMARY"):
-            summary_path = Path(summary_path).resolve()
-            with summary_path.open("a") as f:
-                f.write("## Failed tests\n")
-                f.write("Run the following command to run the failed tests:\n")
-                f.write("```bash\n")
-                f.write(" ".join(["pytest"] + ids) + "\n")
-                f.write("```\n\n")
-
-        # If some tests failed at installing mlflow, we suggest using `--serve-wheel` flag.
-        # Some test cases try to install mlflow via pip e.g. model loading. They pins
-        # mlflow version to install based on local environment i.e. dev version ahead of
-        # the latest release, hence it's not found on PyPI. `--serve-wheel` flag was
-        # introduced to resolve this issue, which starts local PyPI server and serve
-        # an mlflow wheel based on local source code.
-        # Ref: https://github.com/mlflow/mlflow/pull/10247
-        msg = f"No matching distribution found for mlflow=={VERSION}"
-        for rep in failed_test_reports:
-            if any(msg in t for t in (rep.longreprtext, rep.capstdout, rep.capstderr)):
-                terminalreporter.section("HINTS", yellow=True)
-                terminalreporter.write(
-                    f"Found test(s) that failed with {msg!r}. Adding"
-                    " --serve-wheel` flag to your pytest command may help.\n\n",
-                    yellow=True,
-                )
-                break
-
-    main_thread = threading.main_thread()
-    if threads := [t for t in threading.enumerate() if t is not main_thread]:
-        terminalreporter.section("Remaining threads", yellow=True)
-        for idx, thread in enumerate(threads, start=1):
-            terminalreporter.write(f"{idx}: {thread}\n")
-
-        # Uncomment this block to print tracebacks of non-daemon threads
-        # if non_daemon_threads := [t for t in threads if not t.daemon]:
-        #     frames = sys._current_frames()
-        #     terminalreporter.section("Tracebacks of non-daemon threads", yellow=True)
-        #     for thread in non_daemon_threads:
-        #         thread.join(timeout=1)
-        #         if thread.is_alive() and (frame := frames.get(thread.ident)):
-        #             terminalreporter.section(repr(thread), sep="~")
-        #             terminalreporter.write("".join(traceback.format_stack(frame)))
-
-    try:
-        import psutil
-    except ImportError:
-        pass
-    else:
-        current_process = psutil.Process()
-        if children := current_process.children(recursive=True):
-            terminalreporter.section("Remaining child processes", yellow=True)
-            for idx, child in enumerate(children, start=1):
-                terminalreporter.write(f"{idx}: {child}\n")
-
-
-@pytest.fixture(scope="module", autouse=not IS_TRACING_SDK_ONLY)
-def clean_up_envs():
-    """
-    Clean up virtualenvs and conda environments created during tests to save disk space.
-    """
-    yield
-
-    if "GITHUB_ACTIONS" in os.environ:
-        from mlflow.utils.virtualenv import _get_mlflow_virtualenv_root
-
-        shutil.rmtree(_get_mlflow_virtualenv_root(), ignore_errors=True)
-        if not is_windows():
-            conda_info = json.loads(subprocess.check_output(["conda", "info", "--json"], text=True))
-            root_prefix = conda_info["root_prefix"]
-            regex = re.compile(r"mlflow-\w{32,}")
-            for env in conda_info["envs"]:
-                if env == root_prefix:
-                    continue
-                if regex.fullmatch(os.path.basename(env)):
-                    shutil.rmtree(env, ignore_errors=True)
-
-
-@pytest.fixture(scope="session", autouse=True)
-def enable_mlflow_testing():
-    with pytest.MonkeyPatch.context() as mp:
-        mp.setenv(_MLFLOW_TESTING.name, "TRUE")
-        yield
-
-
-@pytest.fixture(scope="session", autouse=not IS_TRACING_SDK_ONLY)
-def serve_wheel(request, tmp_path_factory):
-    """
-    Models logged during tests have a dependency on the dev version of MLflow built from
-    source (e.g., mlflow==1.20.0.dev0) and cannot be served because the dev version is not
-    available on PyPI. This fixture serves a wheel for the dev version from a temporary
-    PyPI repository running on localhost and appends the repository URL to the
-    `PIP_EXTRA_INDEX_URL` environment variable to make the wheel available to pip.
-    """
-    from tests.helper_functions import get_safe_port
-
-    if not request.config.getoption("--serve-wheel"):
-        yield  # pytest expects a generator fixture to yield
-        return
-
-    root = tmp_path_factory.mktemp("root")
-    mlflow_dir = root.joinpath("mlflow")
-    mlflow_dir.mkdir()
-    port = get_safe_port()
-    try:
-        repo_root = subprocess.check_output(
-            [
-                "git",
-                "rev-parse",
-                "--show-toplevel",
-            ],
-            text=True,
-        ).strip()
-    except subprocess.CalledProcessError:
-        # Some tests run in a Docker container where git is not installed.
-        # In this case, assume we're in the root of the repo.
-        repo_root = "."
-
-    subprocess.run(
-        [
-            sys.executable,
-            "-m",
-            "pip",
-            "wheel",
-            "--wheel-dir",
-            mlflow_dir,
-            "--no-deps",
-            repo_root,
-        ],
-        check=True,
-    )
-    with subprocess.Popen(
-        [
-            sys.executable,
-            "-m",
-            "http.server",
-            str(port),
-        ],
-        cwd=root,
-    ) as prc:
-        try:
-            url = f"http://localhost:{port}"
-            if existing_url := os.environ.get("PIP_EXTRA_INDEX_URL"):
-                url = f"{existing_url} {url}"
-            os.environ["PIP_EXTRA_INDEX_URL"] = url
-            # Set the `UV_INDEX` environment variable to allow fetching the wheel from the
-            # url when using `uv` as environment manager
-            os.environ["UV_INDEX"] = f"mlflow={url}"
-            yield
-        finally:
-            prc.terminate()
